% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{adj_rho_sqrd}
\alias{adj_rho_sqrd}
\title{Calculate Adjusted Rho Squared}
\usage{
adj_rho_sqrd(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate Adjusted Rho Squared
}
