% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{aic}
\alias{aic}
\title{Calculate AIC}
\usage{
aic(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate AIC
}
