% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_choices.R
\name{analyze_choices}
\alias{analyze_choices}
\title{Analyzes choices}
\usage{
analyze_choices(db, validated_options)
}
\arguments{
\item{db}{A data.frame containing the estimation data}

\item{validated_options}{A list of validated model options returned from 
\code{\link{validate}}}
}
\value{
A matrix with the choice shares and test statistics for difference in
means
}
\description{
The function compares the market shares across sub-samples in the data. The 
user can specify which, if any, explanators to use in the choice analysis. 
The choice analysis does not involve any estimation.
}
\details{
The function is a rewritten version of \code{apollo_choiceAnalysis()}. See
references.
}
\examples{
\dontrun{
  model_opt <- list(
    choice_analysis_explanators = NULL
  )
  
  # OR
  
  model_opt <- list(
    choice_analysis_explanators = c("male", "income")
  )
  
  analyze_choices(db, model_opt)
}

}
\references{
Hess, S. & Palma, D. (2019), Apollo: a flexible, powerful and customisable 
freeware package for choice model estimation and application, Journal of 
Choice Modelling, Volume 32, September 2019, 100170
Hess, S. & Palma, D. (2019), Apollo version 0.1.0, user manual,
 www.ApolloChoiceModelling.com
}
