% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_objects.R
\name{attach_objects}
\alias{attach_objects}
\title{Attaches the objects passed through the list}
\usage{
attach_objects(objects)
}
\arguments{
\item{objects}{An un-named list of objects to be attached}
}
\value{
Nothing
}
\description{
This function attaches all objects passed through the list \code{objects}.
Each list element is an object to be attached.
}
