% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R
\name{augment.cmdlr}
\alias{augment.cmdlr}
\title{Augment a cmdlr object}
\usage{
\method{augment}{cmdlr}(x, ...)
}
\arguments{
\item{x}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to glance}
}
\description{
Returns the predicted probabilities in a tidy format
}
