% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_draws.R
\name{avg_draws}
\alias{avg_draws}
\title{Average over draws}
\usage{
avg_draws(pr_seq, N)
}
\arguments{
\item{pr_seq}{The probability of the sequence returned from
\code{panel_product}}

\item{N}{The number of individuals in the sample}
}
\value{
A vector of likelihoods
}
\description{
Average over draws for the same respondent
}
