% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{bic_star}
\alias{bic_star}
\title{Calculate BIC*}
\usage{
bic_star(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate BIC*
}
