% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-vcov.R
\name{bread}
\alias{bread}
\title{Create the bread of the sandwich}
\usage{
bread(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Create the bread of the sandwich
}
