% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{caic_star}
\alias{caic_star}
\title{Calculate CAIC*}
\usage{
caic_star(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate CAIC*
}
