% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{cat_summary_info}
\alias{cat_summary_info}
\title{Wrapper for cat sequence}
\usage{
cat_summary_info(x, ...)
}
\arguments{
\item{x}{A 'cmdlr' model object}

\item{...}{Other arguments passed to the function}
}
\description{
Wrapper for cat sequence
}
