% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Function to check the data}
\usage{
check_data(db, estim_opt, model_opt)
}
\arguments{
\item{db}{Data}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}
}
\value{
A list of conditions
}
\description{
The function will check the data to see if it is balanced. It will return
a list of conditions indicating which aspects of the data is unbalanced,
which enables the code to call the correct functions to set up the data.
}
