% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmldR.R
\docType{package}
\name{cmldR}
\alias{cmldR}
\alias{cmldR-package}
\title{cmdlR: Choice modeling in R}
\description{
cmldR is a package for choice modeling in R. It contains functions to handle 
data input, setup for multicor estimation and several pre-programmed 
likelihood functions that will quickly get you analyzing data.
}
