% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cmdlr}
\alias{coef.cmdlr}
\title{S3 generic for coef}
\usage{
\method{coef}{cmdlr}(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Other arguments passed to the function}
}
\description{
S3 generic for coef
}
