% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate.R
\name{collate}
\alias{collate}
\title{Collate different model objects into a single .csv file}
\usage{
collate(path = NULL, pattern = NULL, digits = 4)
}
\arguments{
\item{path}{The relative path (from working directory) to the folder where 
the outputs are stored. Default is NULL, which means outputs are found in 
the root working directory}

\item{pattern}{A regular expression. Only file names which match the regular
expression will be returned.}

\item{digits}{An integer giving the number of digits to round results}
}
\description{
The function looks in the output folder and collects all model objects given
by the \code{pattern} and collates them into a single .csv file for easy
comparison. To make comparison easier, the function will create a vector of
unique parameter names and match parameters from all models to the vector.
This ensures that the same parameters from different models are in the same
row in the output file.
}
