% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_choices.R
\name{compare_choice_shares}
\alias{compare_choice_shares}
\title{Compare choice shares}
\usage{
compare_choice_shares(db, str_choice, alt_avail, explanators = NULL)
}
\arguments{
\item{db}{A data.frame or tibble}

\item{str_choice}{A character string giving the name of the choice variable}

\item{alt_avail}{A list the length of the total number of alternatives. Each
list element is a vector of length equal to the number of observations with
the values 0 and 1 indicating unvailable/available aleternatives in each 
choice situation}

\item{explanators}{A character vector with the names of the variables for
which to calcualte differences in choice shares

The function is a rewritten version of \code{apollo_choiceAnalysis()}.}
}
\value{
A list with the choice shares and test statistics for difference in
means
}
\description{
The function calculates the average value of the explanator for each
alternative when chosen and not chosen.
}
\references{
Hess, S. & Palma, D. (2019), Apollo: a flexible, powerful and customisable 
freeware package for choice model estimation and application, Journal of 
Choice Modelling, Volume 32, September 2019, 100170
Hess, S. & Palma, D. (2019), Apollo version 0.1.0, user manual,
 www.ApolloChoiceModelling.com
}
