% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{convergence_criteria}
\alias{convergence_criteria}
\title{Calculate the convergence criteria}
\usage{
convergence_criteria(object, robust = TRUE, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{robust}{A boolean equal to TRUE if you want to return a robust 
variance covariance matrix based on a simple sandwich estimator}

\item{...}{Other arguments passed to the function}
}
\value{
A named numeric
}
\description{
Calculate the convergence criteria
}
