% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coral.R
\docType{data}
\name{data_coral}
\alias{data_coral}
\title{A stated choice experiment on cold-water coral protection in Norway}
\format{A dataframe in the "wide" format with 4439 choice observations and 31
variables. Each respondent answered up to 12 choice tasks described by
2 experimentally designed alternatives and the SQ. 

\describe{
 \item{id}{Individual ID variable}
 \item{ct}{Choice task ID variable}
 \item{choice_1}{Choice indicator}
 \item{size_1}{Size of the protected area (attribute 1)}
 \item{small_1}{Indicator variable equal to 1 for a small increase (attribute 1)}
 \item{large_1}{ Indicator variable equal to 1 for a large increase (attribute 1)}
 \item{oil_1}{Indicator variable equal to 1 if the proposed area is important to the oil and gas industry (attribute 2) }
 \item{fish_1}{Indicator variable equal to 1 if the proposed area is important to the fisheries (attribute 2)}
 \item{habitat_1}{Indicator variable equal to 1 if the proposed area is important habitat for fish (attribute 3)}
 \item{cost_1}{scaled cost variable measured as a lump sum increase in annual taxes (attribute 4) }
 \item{choice_2}{Choice indicator}
 \item{size_2}{Size of the protected area (attribute 1)}
 \item{small_2}{Indicator variable equal to 1 for a small increase (attribute 1)}
 \item{large_2}{ Indicator variable equal to 1 for a large increase (attribute 1)}
 \item{oil_2}{Indicator variable equal to 1 if the proposed area is important to the oil and gas industry (attribute 2) }
 \item{fish_2}{Indicator variable equal to 1 if the proposed area is important to the fisheries (attribute 2)}
 \item{habitat_2}{Indicator variable equal to 1 if the proposed area is important habitat for fish (attribute 3)}
 \item{cost_2}{scaled cost variable measured as a lump sum increase in annual taxes (attribute 4) }
 \item{choice_3}{Choice indicator}
 \item{size_3}{Size of the protected area (attribute 1)}
 \item{small_3}{Indicator variable equal to 1 for a small increase (attribute 1)}
 \item{large_3}{ Indicator variable equal to 1 for a large increase (attribute 1)}
 \item{oil_3}{Indicator variable equal to 1 if the proposed area is important to the oil and gas industry (attribute 2) }
 \item{fish_3}{Indicator variable equal to 1 if the proposed area is important to the fisheries (attribute 2)}
 \item{habitat_3}{Indicator variable equal to 1 if the proposed area is important habitat for fish (attribute 3)}
 \item{cost_3}{scaled cost variable measured as a lump sum increase in annual taxes (attribute 4) }
 \item{male}{An indicator equal to 1 if a respondent was male}
 \item{university}{An indicator equal to 1 if a respondent has university education}
 \item{hh_income}{An interval variable for houshold income. 1 = Below NOK 200K, 2 = NOK 200 - 400K, 3 = NOK 400 - 600K, 4 = NOK 600 - 800K, 5 = NOK 800 - 1000K, 6 = NOK 1.0 - 1.2 Million, 7 = NOK 1.2 - 1.5 Million, 8 = Above NOK 1.5 Million}
 \item{adults}{Number of people above 18 in the household}
 \item{children}{Number of people below 18 in the household}
}}
\source{
\url{https://doi.org/10.18710/Q39V7O}
}
\usage{
data_coral
}
\description{
This data is from a stated choice discrete choice experiment on cold-water
coral protection in Norway. The data was gathered using 22 valuation 
workshops in select municipalities in Norway.
}
\references{
Aanesen, M., Armstrong. C., Czajkowski, M., Falk-Petersen, J., Hanley, N. & Navrud, S., 2015, Willingness-to-pay for unfamiliar public goods: Preserving cold-water coral in Norway, Ecological Economics 112, 53-67, https://doi.org/10.1016/j.ecolecon.2015.02.007
Sandorf, E.D., Aanesen, M. & Navrud, S., 2016, Valuing unfamiliar and complex environmental goods: A comparison of valuation workshops and internet panel surveys with videos, Ecological Economics 129, 50-61
Sandorf, E.D., Campbell, D. & Hanley, N., 2017, Disentangling the influence of knowledge on attribute non-attendance, Journal of Choice Modelling, 24, 36-50
}
\keyword{datasets}
