% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{dbic}
\alias{dbic}
\title{Calculate DBIC}
\usage{
dbic(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate DBIC
}
