% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detach_objects.R
\name{detach_objects}
\alias{detach_objects}
\title{Detach objects from the environment}
\usage{
detach_objects(objects = NULL, all = FALSE)
}
\arguments{
\item{objects}{An un-named list of objects to be attached}

\item{all}{If \code{TRUE} the function detaches all non-essential objects and
packages from the global environment.}
}
\value{
Nothing
}
\description{
This function detaches the objects passed through the list \code{objects}. 
Each list element is an object to be detached.
}
