% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-helpers.R
\name{difference_utility}
\alias{difference_utility}
\title{Take the difference between all utilities and the chosen alternative}
\usage{
difference_utility(V, v_chosen)
}
\arguments{
\item{V}{A list of utilities}

\item{v_chosen}{A vector of the chosen utilities}
}
\description{
Take the difference between all utilities and the chosen alternative
}
