% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimates the model}
\usage{
estimate(inputs)
}
\arguments{
\item{inputs}{List of inputs to the estimation. The result of running
\code{prepare}.}
}
\description{
The function is a wrapper for several cmdlR functions that sets up parallel
processes and estimates the data.
}
