% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-model.R
\name{estimate_bgw}
\alias{estimate_bgw}
\title{Estimate the model using the 'bgw' package}
\usage{
estimate_bgw(log_lik, param_free, control)
}
\arguments{
\item{log_lik}{A log-likelihood expression prepared by 
\code{\link{prepare_log_lik}}}

\item{param_free}{A list of freely estimated parameters}

\item{control}{A list of estimation controls passed to the optimizer}
}
\value{
A custom model object with the results of the estimation
}
\description{
The function is a convenient wrapper around \code{\link{bgw_mle}}.
}
