% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-model.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate the model}
\usage{
estimate_model(
  type,
  log_lik,
  param_start,
  param_free,
  param_fixed,
  workers,
  ll,
  control
)
}
\arguments{
\item{type}{A character string giving the R pacakge to use for optimization.
One of: "maxlik" or "ucminf"}

\item{log_lik}{A log-likelihood expression prepared by 
\code{\link{prepare_log_lik}}}

\item{param_start}{The complete list of parameters at their starting values
(including fixed parameters)}

\item{param_free}{A list of freely estimated parameters}

\item{param_fixed}{A list of parameters that are held fixed at their 
starting values}

\item{workers}{A list of workers for parallel computing}

\item{ll}{The 'raw' log likelihood function passed by the user}

\item{control}{A list of estimation controls passed to the optimizer}
}
\value{
A custom model object
}
\description{
The function is a wrapper around several methods for obtaining the optimum
using either maximization (maxLik) or minimization (ucminf). The function 
is meant for internal use and is not exported.
}
