% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_choices.R
\name{expand_choices}
\alias{expand_choices}
\title{Expand choice variable}
\usage{
expand_choices(db, str_id, str_ct, str_choice)
}
\arguments{
\item{db}{A data.frame or tibble}

\item{str_id}{A character string giving the name of the individual id 
variable}

\item{str_ct}{A character string giving the name of the choice task variable}

\item{str_choice}{A character string giving the name of the choice variable}
}
\value{
A tibble with the expanded choice variable
}
\description{
Expands the vector of choice variables to a tibble with dummy columns for
whether an alternative was chosen. The function uses the sorted unique
values of the choice variable to infer the total number of alternatives
available (irrespective of whether any of them were chosen)
}
