% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{final_param}
\alias{final_param}
\title{Extract final parameters from the model object}
\usage{
final_param(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract final parameters from the model object
}
