% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{fixed_param}
\alias{fixed_param}
\title{Extract fixed parameters from the model object}
\usage{
fixed_param(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract fixed parameters from the model object
}
