% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_1}
\alias{fn_1}
\title{Helper function fn_1}
\usage{
fn_1(a, L)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{L}{The lower cholesky matrix}
}
\description{
Helper function to calculate the standard errors of standard deviations and 
correlations from a MIXL model. Together, all the helper functions and the 
four functions \code{\link{fn_se_var}}, \code{\link{fn_se_cov}}, 
\code{\link{fn_se_corr}} and \code{\link{fn_se_std}}, is an application 
of the Delta method. The helper functions are for internal use only and are 
not exported.
}
\references{
http://www.stephanehess.me.uk/software.html
Daly, A. J., Hess, S., & de Jong, G., 2012, Calculating errors for measures derived
from choice modelling estimates, Transportation Research Part B, 46(2):333-341
}
