% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_1d}
\alias{fn_1d}
\title{Helper function fn_1d}
\usage{
fn_1d(a, j, k, L)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{j}{Indexing parameter}

\item{k}{Indexing parameter}

\item{L}{The lower cholesky matrix}
}
\description{
For a description see \code{\link{fn_1}}
}
