% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_2dv}
\alias{fn_2dv}
\title{Helper function fn_2dv}
\usage{
fn_2dv(a, b, L)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{b}{An integer indicating the column position of the random parameter in the
asymptotic variance-covariance matrix}

\item{L}{The lower cholesky matrix}
}
\description{
For a description see \code{\link{fn_1}}
}
