% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_4dv}
\alias{fn_4dv}
\title{Helper function fn_4dv}
\usage{
fn_4dv(a, L)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{L}{The lower cholesky matrix}
}
\description{
For a description see \code{\link{fn_1}}
}
