% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_se_std}
\alias{fn_se_std}
\title{Calcualte the standard error of the standard deviation}
\usage{
fn_se_std(a, L, lambda)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{L}{The lower cholesky matrix}

\item{lambda}{The asymptotic variance covariance matrix}
}
\value{
A vector of standard errors
}
\description{
The standard deviation is calculated as the square root of the diagonal of 
the variance coavariance matrix of your random parameters. The variance-
covariance matrix is calculated using the estimated lower Cholesky matrix.
}
\references{
http://www.stephanehess.me.uk/software.html
Daly, A. J., Hess, S., & de Jong, G., 2012, Calculating errors for measures derived
from choice modelling estimates, Transportation Research Part B, 46(2):333-341
}
