% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_se.R
\name{fn_se_var}
\alias{fn_se_var}
\title{Calculate the standard error of the variance}
\usage{
fn_se_var(a, L, lambda)
}
\arguments{
\item{a}{An integer indicating the row position of the random parameter in the
asymptotic variance-covariance matrix}

\item{L}{The lower cholesky matrix}

\item{lambda}{The asymptotic variance covariance matrix}
}
\description{
Calculate the standard error of the variance terms in the variance-covariance
matrix created from the lower Cholesky decomposition matrix of the random 
parameters in a mixed logit model.
}
\references{
http://www.stephanehess.me.uk/software.html
Daly, A. J., Hess, S., & de Jong, G., 2012, Calculating errors for measures derived
from choice modelling estimates, Transportation Research Part B, 46(2):333-341
}
