% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{free_param}
\alias{free_param}
\title{Extract free parameters from the model object}
\usage{
free_param(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract free parameters from the model object
}
