% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{function_value_zero}
\alias{function_value_zero}
\title{Extract the function value at zero from the model object}
\usage{
function_value_zero(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\value{
A single numeric value
}
\description{
Extract the function value at zero from the model object
}
