% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{function_values}
\alias{function_values}
\title{Extract the function values}
\usage{
function_values(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional arguments passed to the function}
}
\value{
A vector of function values at the optimum
}
\description{
Extract the function values
}
