% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_control}
\alias{get_control}
\title{Extract the control list}
\usage{
get_control(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the control list
}
