% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_convergence}
\alias{get_convergence}
\title{Extract the convergence of the model}
\usage{
get_convergence(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the convergence of the model
}
