% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_convergence_code}
\alias{get_convergence_code}
\title{Extract the convergence code}
\usage{
get_convergence_code(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the convergence code
}
