% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_convergence_message}
\alias{get_convergence_message}
\title{Extract the convergence message}
\usage{
get_convergence_message(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the convergence message
}
