% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_description}
\alias{get_description}
\title{Extract the description of the model}
\usage{
get_description(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the description of the model
}
