% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_estimation_end}
\alias{get_estimation_end}
\title{Extract end time for estimation}
\usage{
get_estimation_end(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract end time for estimation
}
