% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_estimation_start}
\alias{get_estimation_start}
\title{Extract start time for estimation}
\usage{
get_estimation_start(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract start time for estimation
}
