% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_function_value}
\alias{get_function_value}
\title{Extract the function value from the model object}
\usage{
get_function_value(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\value{
A single numeric value
}
\description{
Extract the function value from the model object
}
