% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_function_values}
\alias{get_function_values}
\title{Extract the function values}
\usage{
get_function_values(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\value{
A vector of function values at the optimum
}
\description{
Extract the function values
}
