% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_gradient}
\alias{get_gradient}
\title{Extract the gradient of a model object}
\usage{
get_gradient(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the gradient of a model object
}
