% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_gradient_obs}
\alias{get_gradient_obs}
\title{Extract the gradient observations from a model object}
\usage{
get_gradient_obs(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the gradient observations from a model object
}
