% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_hessian}
\alias{get_hessian}
\title{Extract the Hessian}
\usage{
get_hessian(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the Hessian
}
