% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_model_frame}
\alias{get_model_frame}
\title{Extract the model_frame}
\usage{
get_model_frame(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the model_frame
}
