% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_name}
\alias{get_name}
\title{Extract the name of the model}
\usage{
get_name(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the name of the model
}
