% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_param_final}
\alias{get_param_final}
\title{Extract final parameters from the model object}
\usage{
get_param_final(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract final parameters from the model object
}
