% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_param_free}
\alias{get_param_free}
\title{Extract free parameters from the model object}
\usage{
get_param_free(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract free parameters from the model object
}
