% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_param_start}
\alias{get_param_start}
\title{Extract starting parameters from the model object}
\usage{
get_param_start(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract starting parameters from the model object
}
