% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-helpers.R
\name{get_v_chosen}
\alias{get_v_chosen}
\title{Calculate the utility of the chosen alternative}
\usage{
get_v_chosen(V, choice_var)
}
\arguments{
\item{V}{A list of utilities}

\item{choice_var}{A a numeric vector with the chosen alternative by choice
task}
}
\value{
A vector with the utility of the chosen alternative
}
\description{
Calculate the utility of the chosen alternative
}
