% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_info.R
\name{get_worker_info}
\alias{get_worker_info}
\title{Function for getting worker info}
\usage{
get_worker_info(workers)
}
\arguments{
\item{workers}{A cluster of workers}
}
\description{
The function will loop through the workers to get information about what
objects are loaded on each worker and how the work load is distributed.
}
