% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{gradient}
\alias{gradient}
\title{Extract the gradient of a model object}
\usage{
gradient(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract the gradient of a model object
}
