% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{hessian_complete}
\alias{hessian_complete}
\title{Check if the Hessian matrix is complete}
\usage{
hessian_complete(hess)
}
\arguments{
\item{hess}{A hessian matrix}
}
\value{
A boolean equal to TRUE if the Hessian matrix could not be calculated
or contains missing values/NaN
}
\description{
Check if the Hessian matrix is complete
}
