% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{hessian_condition_number}
\alias{hessian_condition_number}
\title{Get the hessian condition number}
\usage{
hessian_condition_number(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Other arguments passed to the function}
}
\description{
The condition number is the ratio of the largest to the smallest Eigen value
of the Hessian matrix. If this number is very large, then the Hessian matrix
is ill-conditioned. If this number is smaller than machine precision, the
Hessian is numerically rank deficient as is 'indefinite'
}
