% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{hqic}
\alias{hqic}
\title{Calculate HQIC}
\usage{
hqic(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate HQIC
}
