% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{ht_aic}
\alias{ht_aic}
\title{Calculate HTAIC}
\usage{
ht_aic(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Calculate HTAIC
}
