% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_choices.R
\name{identify_choice_patterns}
\alias{identify_choice_patterns}
\title{Identify choice patterns}
\usage{
identify_choice_patterns(db, str_id, str_ct, str_choice, sq_alt)
}
\arguments{
\item{db}{A data.frame or tibble}

\item{str_id}{A character string giving the name of the individual id 
variable}

\item{str_ct}{A character string giving the name of the choice task variable}

\item{str_choice}{A character string giving the name of the choice variable}

\item{sq_alt}{A number giving the alternative (in the choice variable) 
corresponding to the status quo or opt out}
}
\value{
A vector the length of rows in `db` with value equal to 1 if an 
individual always chose the SQ or opt out alternative
}
\description{
Identify choice patterns
}
