% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inspect_list}
\alias{inspect_list}
\title{Inspect a list}
\usage{
inspect_list(x)
}
\arguments{
\item{x}{A list of vectors or matrices}
}
\description{
A useful function for exploratory work or when coding. Takes a list as an
argument and prints a glimpse of the data contained in each list element 
and the dimensions of each list element. If the list contains more than 5
list elements, only the first 5 elements are printed to console with no 
warning. The function will print a maximum of 10 values if the list element
is a vector and 10 rows and 5 columns if the list element is a matrix.
}
\details{
NOTE: There are currently no checks built in to handle a list of lists. In
this case, everything in the sublists will be printed with no checks or 
warnings.
}
\examples{
x <- list(
  matrix(runif(100), nrow = 20, ncol = 5),
  matrix(runif(100), nrow = 10, ncol = 10)
)

inspect_list(x)

}
