% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc_probs.R
\name{lc_unconditional}
\alias{lc_unconditional}
\title{Calculate the unconditional probabilities}
\usage{
lc_unconditional()
}
\value{
The matrix of unconditional class probabilities
}
\description{
Calculate the unconditional probabilities
}
