% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{make_model_name}
\alias{make_model_name}
\title{Create the model name used when saving the data}
\usage{
make_model_name(x)
}
\arguments{
\item{x}{A string}
}
\description{
Create the model name used when saving the data
}
