% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-vcov.R
\name{meat}
\alias{meat}
\title{Create the meat of the sandwich}
\usage{
meat(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
Create the meat of the sandwich
}
