% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_probs.R
\name{mnl_probabilities}
\alias{mnl_probabilities}
\title{Calculate the MNL probability}
\usage{
mnl_probabilities(V, choice_var, alt_avail)
}
\arguments{
\item{V}{A list of utilities}

\item{choice_var}{A vector with the chosen alternative}

\item{alt_avail}{A list of alternative availabilities}
}
\description{
Function is convenient way to calculate the MNL probabilities.
}
\details{
Function for calculating the MNL probabilities

Calculate the MNL probabilities taking the difference from the chosen
alternative.
}
