% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_probs.R
\name{mnl_probs}
\alias{mnl_probs}
\title{Function for calculating the MNL probabilities}
\usage{
mnl_probs(model_opt)
}
\arguments{
\item{model_opt}{List of model options}
}
\description{
Calculate the MNL probabilities taking the difference from the chosen
alternative.
}
