% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_probs.R
\name{mnl_zero}
\alias{mnl_zero}
\title{Calculate the probability at zero}
\usage{
mnl_zero(alt_avail)
}
\arguments{
\item{alt_avail}{List of alternative availabilities}
}
\value{
A vector of length equal to the number of choice observations in 
the data (incl. NAs from the padding during data setup)
}
\description{
Calculates the probability at zero parameter values only across the 
available alternatives.
}
