% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Calculate model fit}
\usage{
model_fit(object, type = "bic", ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{type}{A character string indicating the fit statistic to calculate. Has
to be one of: "adj_rho_sqrd", "aic", "aic3", "caic", "caic_star", "ht_aic",
bic", "bic_star", "dbic" or "hqic". The default is "bic".}

\item{...}{Additional parameters passed to the function}
}
\value{
A named numeric
}
\description{
Calculate model fit
}
