% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs}
\alias{nobs}
\title{S3 generic for nobs}
\usage{
nobs(object, use.fallback = FALSE, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{use.fallback}{Should fallback methods be used to try to guess the
value?}

\item{...}{Additional arguments passed}
}
\description{
A generic method for getting the number of observations based on the number
of rows in the matrix of gradient observations.
}
