% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{normal_vcov}
\alias{normal_vcov}
\title{Normal variance-covariance matrix}
\usage{
normal_vcov(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Other arguments passed to the function}
}
\value{
A matrix with row and column names equal to the parameters of the
fitted model
}
\description{
The function is a wrapper around the [MASS::] from the MASS package
to calculate the (negative) inverse of the Hessian matrix.
}
