% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_deriv.R
\name{num_deriv}
\alias{num_deriv}
\title{Obtain the numerical derivative by central difference}
\usage{
num_deriv(param, log_lik, N, eps)
}
\arguments{
\item{param}{A vector of parameters}

\item{log_lik}{The log-likelihood function}

\item{N}{The number of observations, e.g. individuals}

\item{eps}{The precision level}
}
\value{
A matrix with rows equal to N and columns equal to the number of
parameters
}
\description{
The function calculates the numerical derivate of an objective function using
central difference. It is a relatively quick way to obtain the score
contribution of each observation when the optimizer requires a single value.
}
