% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_data.R
\name{pad_data}
\alias{pad_data}
\title{Pads the data}
\usage{
pad_data(db, model_opt)
}
\arguments{
\item{db}{Data}

\item{model_opt}{List of model options}
}
\value{
A tibble with the padded data.
}
\description{
The function pads the data by adding rows with NA in case of missing choice
occasions.
}
