% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{pairwise_chisq_test}
\alias{pairwise_chisq_test}
\title{A pairwise Chi-squared test}
\source{
Modified from here: https://stats.stackexchange.com/questions/85664/r-procedure-for-comparing-multiple-categorical-variables-similar-to-anova-fol
}
\usage{
pairwise_chisq_test(x, g, p.adjust.method = p.adjust.methods, ...)
}
\arguments{
\item{x}{A response vector}

\item{g}{A grouping vector. If not a factor it will be converted to one.}

\item{p.adjust.method}{A character string specifying the method for multiple
testing adjustments. See \link{\code{p.adjust.methods}}}

\item{...}{Additional arguments passed to \link{\code{chisq.test}}}
}
\description{
A pairwise Chi-squared test
}
