% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_product.R
\name{panel_product}
\alias{panel_product}
\title{Take the panel product}
\usage{
panel_product(pr_chosen, S)
}
\arguments{
\item{pr_chosen}{The probability of the chosen alternative returned from
\code{mnl_probs}}

\item{S}{The number of choie occasions per individual}
}
\value{
A vector that is N * DRAWS stacked by
}
\description{
Takes the product across the same choice situations for the same individual.
}
