% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{parameter_info}
\alias{parameter_info}
\title{Summary of parameter info}
\usage{
parameter_info(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\value{
A tidy [tibble::tibble()] summarizing the parameter information
}
\description{
Creating a summary of parameter information including starting values, 
final values, and Eigen values of the Hessian matrix.
}
