% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_choices.R
\name{plot_choice}
\alias{plot_choice}
\title{Plot choice shares}
\usage{
plot_choice(db, str_id, str_ct, str_choice)
}
\arguments{
\item{db}{A data.frame or tibble}

\item{str_id}{A character string giving the name of the individual id 
variable}

\item{str_ct}{A character string giving the name of the choice task variable}

\item{str_choice}{A character string giving the name of the choice variable}
}
\value{
A plot
}
\description{
Plot choice shares
}
