% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predicts "the future"}
\usage{
predict(model)
}
\arguments{
\item{model}{A model opbject returned by \code{estimate}}
}
\description{
This function is a wrapper for several functions that allow you to use your
model results to make predictions.
}
