% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_alt_avail}
\alias{prepare_alt_avail}
\title{Prepare the alternative availabilities}
\usage{
prepare_alt_avail(alt_avail, db)
}
\arguments{
\item{alt_avail}{A list of alternative availabilites supplied by the user}

\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}
}
\value{
A list of alternative availabilities
}
\description{
Sets up the list of alternative availabilities. If alternative availability
is an integer equal to 1, i.e. always available, then repeats it equal to 
the number of rows in the (padded) data, if not includes the actual 
availability vector.
}
