% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to check the data}
\usage{
prepare_data(db, estim_opt, model_opt)
}
\arguments{
\item{db}{Data}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}
}
\value{
A list of conditions
}
\description{
The function will check the data to see if it is balanced. It will take the
necessary steps to balance the data. If it is unable to do so, the code will
return suggestions for how to set up the data.
}
