% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_draws}
\alias{prepare_draws}
\title{Prepares the draws}
\usage{
prepare_draws(n_id, n_ct, n_draws, type, rpar, seed)
}
\arguments{
\item{n_id}{An integer giving the nubmer of respondents}

\item{n_ct}{An integer giving the number of choice situations}

\item{n_draws}{An integer giving the number of draws per individual}

\item{type}{A character string giving the type of draws to use}

\item{rpar}{A named list giving the type of distribution}

\item{seed}{An integer giving the seed to change the scrambling of the sobol
sequence}
}
\value{
A list of matrixes matching 'rpar' where each matrix has the 
dimensions n_id*n_ct x n_draws
}
\description{
In a model using simulation, we need to prepare a set of random draws to 
use when simulating/approximating the distributions/integrals. The function
is a convenient wrapper around the function \code{\link{make_draws}}.
}
\details{
Depending on the model options, the function will transform the supplied
uniform draws to either standard uniform, standard normal or standard
triangular. It is up to the user to specify the correct distributions in the
log-likelihood function.
}
