% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_estim_env.R
\name{prepare_estim_env}
\alias{prepare_estim_env}
\title{Prepare the estimation environment}
\usage{
prepare_estim_env(db, draws, model_opt)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} with the data to be used in
estimation.}

\item{draws}{A list of draws where each list element correspond to one dimension
or if we are using multiple cores, a list of lists where the top-level list
corresponds to the cores and the lower-level list corresponds to the draws.}

\item{model_opt}{A list of model options including random parameters, draws,
fixed and non-fixed parameters during estimation, starting value search 
procedures etc.}
}
\value{
An environment containing the data, draws and relevant indices to use
in estimation.
}
\description{
To allow variables and parameters to be called by name, we need to set up 
an estimation environment to provide the context in which we evaluate the 
log-likelihood expression.
}
\details{
This function is intended for internal use only.
}
