% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_estimation_environment}
\alias{prepare_estimation_environment}
\title{Prepare the estimation environment}
\usage{
prepare_estimation_environment(
  db,
  alt_avail,
  draws,
  cores,
  str_id,
  str_ct,
  str_choice,
  n_obs,
  param_fixed,
  ...
)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}

\item{alt_avail}{A list of alternative availabilites supplied by the user}

\item{draws}{A list of matrixes with draws for each random parameter in the
model}

\item{cores}{An integer indicating the number of cores to use in estimation}

\item{str_id}{A character string giving the name of the id variable}

\item{str_ct}{A character string giving the name of the ct variable}

\item{str_choice}{A character string giving the name of the choice variable}

\item{n_obs}{An integer giving the number of observations}

\item{param_fixed}{A list of fixed parameters to add to environment}

\item{...}{Additional objects passed on to the estimation environment. NOTE:
these are NOT split by core or respondents, but exported in its entirety.}
}
\value{
An environment or list of environments providing the context for 
evaluating the log-likelihood function.
}
\description{
The estimation environment provides the context for evaluating the
log-likelihood function. Creating an environment allow us to call parameters,
variables and indexes directly without calls to \code{\link{attach}} and 
\code{\link{detach}}.
}
\details{
The use of an estimation environment also makes debugging easy because it is
possible attach the environment directly to run through the
log-likelihood function line-by-line. To efficiently debug, it is 
important that the number of cores is set to 1.

To run the model on subsets of the data, simply pass a subset of the data
using the db argument.
}
