% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parallel.R
\name{prepare_parallel}
\alias{prepare_parallel}
\title{Function to prepare and set up for parallel estimation}
\usage{
prepare_parallel(estim_opt, save_opt)
}
\arguments{
\item{estim_opt}{List of estimation options}

\item{save_opt}{List of options for saving outputs}
}
\description{
The function will set up the workers, split the data and export necessary
objects and functions to the workers.
}
