% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_workers.R
\name{prepare_workers}
\alias{prepare_workers}
\title{Function to prepare the workers for parallel estimation}
\usage{
prepare_workers(db, draws, inputs, workers)
}
\arguments{
\item{db}{Data}

\item{draws}{A list of random draws used in simulation}

\item{inputs}{A list containing options and data to be passed to the worker}

\item{workers}{A cluster of workers}
}
\value{
A cluster of workers
}
\description{
The function will set up the workers, split the data and export necessary
objects and functions to the workers.
}
