% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cmdlr}
\alias{print.cmdlr}
\title{A generic print function for printing a 'cmdlr' object}
\usage{
\method{print}{cmdlr}(x, ...)
}
\arguments{
\item{x}{A 'cmdlr' model object}

\item{...}{Other parameters passed to 'cmdlr'}
}
\description{
A generic print function for printing a 'cmdlr' object
}
