% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R, R/summary.R
\name{print.summary.cmdlr}
\alias{print.summary.cmdlr}
\title{Generic S3 print function}
\usage{
\method{print}{summary.cmdlr}(x, ...)

\method{print}{summary.cmdlr}(x, ...)
}
\arguments{
\item{...}{Other arguments passed to the function}
}
\description{
Generic S3 print function

Generic S3 print function
}
