% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{pval}
\alias{pval}
\title{Calculate p-values}
\usage{
pval(est, std_err, df, hypotheses = NULL)
}
\arguments{
\item{est}{A vector of model estimates}

\item{std_err}{A vector of standard errors}

\item{df}{An integer giving the degrees of freedom for the test}

\item{hypotheses}{A vector of values to test against. The vector must be the
same length as `est`. If no vector is supplied, the default test is against
0.}
}
\value{
A vector of p-values the length of `est`
}
\description{
A convenient function to calculate p-values for **t-statistics**. By default
p-values are calculated based on the null hypothesis that the true parameters
are equal to zero. To test against other values a vector of "hypotheses" 
must be supplied.
}
