% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R, R/methods.R, R/vcov.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{coef}
\alias{nobs}
\alias{vcov}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{stats}{\code{\link[stats]{coef}}, \code{\link[stats]{nobs}}, \code{\link[stats]{vcov}}}
}}

