% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-helpers.R
\name{replace_na_padding}
\alias{replace_na_padding}
\title{Replace the NAs introduced by padding the data}
\usage{
replace_na_padding(x)
}
\arguments{
\item{x}{A matrix or vector of chosen probabilities}
}
\value{
A matrix or vector where NAs are replaced by 1
}
\description{
Replaces the NA values introduced when padding the data. We need to replace
the NAs since no na.remove option exists for the Rfast package. See 
\code{\link{prepare_data}} for details
}
