% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-helpers.R
\name{reshape_panel}
\alias{reshape_panel}
\title{Reshape the panel}
\usage{
reshape_panel(x, nrow)
}
\arguments{
\item{x}{A matrix or vector}

\item{nrow}{Number of rows}
}
\description{
A reshape function relying on R being column major and the data being stacked
and sorted correctly
}
