% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-vcov.R
\name{robust_vcov}
\alias{robust_vcov}
\title{A sandwich estimator for robust standard errors}
\usage{
robust_vcov(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Additional parameters passed to the function}
}
\description{
A sandwich estimator for robust standard errors
}
