% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{Stores the results of the model to files}
\usage{
save(x, path = getwd(), what = "object", ...)
}
\arguments{
\item{x}{A 'cmdlr' model object}

\item{path}{The file path to the folder where the output should be saved. 
The default is the current working directory. The file name is inferred 
based on the model name and the 'what' argument.}

\item{what}{What output to save. One of: 'object', 'summary', 'hessian', 
'vcov'. The default is the model object.}

\item{...}{Additional parameters passed to the function. For example, 
'robust' if you are saving the variance-covariance matrix}
}
\description{
This function takes the model object and the list \code{save_opt} and stores
the specified outputs to files. The model object is stored to an .rds file, 
model summary created using \code{\link{summarize}} is stored to a .txt file
using \code{sink()}, and hessian and variance-covariance matrices are stored
as .csv files. Appropriate checks should be in place to avoid over-writing
existing files and to close open sink connections. However, it is on the user
to be careful with file names to avoid any chance of loosing data.
}
