% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_vcov}
\alias{save_vcov}
\title{Save the variance covariance matrix}
\usage{
save_vcov(x, path, ...)
}
\arguments{
\item{x}{A 'cmdlr' model object}

\item{path}{The file path to the folder where the output should be saved. 
The default is the current working directory. The file name is inferred 
based on the model name and the 'what' argument.}

\item{...}{Additional parameters passed to the function. For example, 
'robust' if you are saving the variance-covariance matrix}
}
\description{
Save the variance covariance matrix
}
