% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_start_values.R
\name{search_simple}
\alias{search_simple}
\title{Simple search}
\usage{
search_simple(
  ll,
  log_lik,
  param_free,
  param_fixed,
  n_candidates,
  n_return,
  multiplier,
  workers
)
}
\arguments{
\item{ll}{This is the 'raw' log-likelihood function passed to the estimation
routine. It is important that the user takes into account whether the 
optimization routine is a minimizer (e.g. 'ucminf') or a maximizer 
(e.g. 'maxlik').}

\item{log_lik}{A log-likelihood expression prepared by 
\code{\link{prepare_log_lik}}}

\item{param_free}{A list of freely estimated parameters}

\item{param_fixed}{A list of parameters that are held fixed at their 
starting values}

\item{n_candidates}{An integer giving the number of candidates to evaluate.
Default is 100.}

\item{n_return}{An integer giving the number of parameter vectors to return.
The default is 10}

\item{multiplier}{A double indicating a multiplier for the 'simple' search
algorithm. The default is 1.}

\item{workers}{A list of workers for parallel computing}
}
\value{
A matrix of starting values (free parameters only)
}
\description{
The simple search algorithm is essentially an evaluation of a large number
of randomly generated vectors of starting values. Each vector is evaluated
at the starting values and sorted in descending order. The best fitting
vectors of starting values are returned.
}
