% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_start_values.R
\name{search_smart}
\alias{search_smart}
\title{Smart search}
\usage{
search_smart(
  ll,
  log_lik,
  param_free,
  param_fixed,
  n_candidates,
  n_return,
  multiplier,
  workers
)
}
\arguments{
\item{ll}{This is the 'raw' log-likelihood function passed to the estimation
routine. It is important that the user takes into account whether the 
optimization routine is a minimizer (e.g. 'ucminf') or a maximizer 
(e.g. 'maxlik').}

\item{log_lik}{A log likelihood expression}

\item{param_free}{A named vector of freely estimated parameters}

\item{param_fixed}{A named vector of parameters to keep fixed at their
starting values during estimation}

\item{n_candidates}{An integer giving the numer of candidates to evaluate.
Default is 100.}

\item{n_return}{An integer giving the number of parameter vectors to return.
The default is 10}

\item{multiplier}{A double indicating a multiplier for the 'simple' search
algorithm. The default is 1.}

\item{workers}{A list of workers or NA if single core}
}
\value{
A matrix of starting values (free parameters only)
}
\description{
Smart search
}
