% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{set_controls}
\alias{set_controls}
\title{Validate and set default estimation controls}
\usage{
set_controls(control)
}
\arguments{
\item{control}{A list of control options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of estimation controls and sets
defaults. The defaults follow those set by the respective optimizer. See
\code{\link{maxLik}} and \code{\link{ucminf}} for details. Additional options
include `optimizer`, `method`, `cores`, `calculate_hessian` and
`calculate_jacobian`.
}
\details{
The function is called once within \code{\link{estimate}}.
}
