% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_estim_opt_defaults.R
\name{set_estim_opt_defaults}
\alias{set_estim_opt_defaults}
\title{Check estimation options}
\usage{
set_estim_opt_defaults(estim_opt_input)
}
\arguments{
\item{estim_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function checks the list of estimation options specified by the user. If
no list of options is supplied or some of the options are not defined, the 
function will supply reasonable default values. The list of estimation options
specifies options related to the optimization routines used.
}
