% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{split_data}
\alias{split_data}
\title{Split the data into a list}
\usage{
split_data(db, str_id, cores)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}

\item{str_id}{A character string giving the name of the id variable}

\item{cores}{An integer indicating the number of cores to use in estimation}
}
\value{
A list of data.frames with length equal to the number of cores to 
use in estimation
}
\description{
The function takes the data and splits it into roughly equal chunks to be
sent to the workers for parallel processing. The function takes care to not
split the same individual across multiple workers.
}
\details{
Note that the function will only split the data correctly if all respondents
have the same number of rows. Therefore, it is important that this function
is called after the data has been padded with NAs in case of unequal 
number of choice tasks.
}
