% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Standardize a variable}
\usage{
standardize(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{A boolean equal to TRUE if ignore NA when calculating the mean
and standard deviations. See \code{\link{mean}} and \code{\link{sd}}}
}
\value{
A vector of standardized numbers
}
\description{
Standardize a variable by subtracting the mean and dividing by the standard
deviation
}
\examples{
x <- runif(10)
standardize(x)

x[1] <- NA
standardize(x, na.rm = TRUE)

}
