% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{stars}
\alias{stars}
\title{Create significance stars}
\usage{
stars(pval)
}
\arguments{
\item{pval}{A vector of p-values}
}
\value{
A character vector of stars indicating significance at different 
levels
}
\description{
A convenient function that creates stars based on any vector of supplied 
p-values.
}
