% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{start_param}
\alias{start_param}
\title{Extract starting parameters from the model object}
\usage{
start_param(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional arguments passed to the function}
}
\description{
Extract starting parameters from the model object
}
