% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store}
\alias{store}
\title{Stores the results of the model to files}
\usage{
store(model, inputs)
}
\arguments{
\item{model}{A model opbject returned by \code{estimate}}

\item{inputs}{A list of model inputs}
}
\description{
This function stores various model outputs to easily accessible .csv files
and .rds objects.
}
