% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{subset_data}
\alias{subset_data}
\title{Subset the data to only include used variables}
\usage{
subset_data(func, db, ...)
}
\arguments{
\item{func}{A user supplied function}

\item{db}{A tibble or data.frame}

\item{...}{Other variables in `db` to keep that are not used in the 
supplied function. Must be supplied as character strings.}
}
\description{
The function deparses the supplied function to find only the variables
used.
}
\examples{
ll <- function(param) {
  V <- list(
    alt1 = asc_1 + b_1 * variable_1,
    alt2 = asc_2 + b_2 * var_2
  )
}

db <- tibble::tibble(
  variable_1 = runif(10),
  var_2 = runif(10),
  id = seq(1, 10, 1),
  ct = rep(1:2, 5),
  income = sample(100:1000, 10)
)

subset_data(ll, db, c("id", "ct"))
subset_data(ll, db, "id", "ct")
subset_data(ll, db, c("id", "ct"), "income")

}
