% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R, R/summary.R
\name{summarize}
\alias{summarize}
\title{Summarizes outputs}
\usage{
summarize(model)

summarize(model)
}
\arguments{
\item{model}{A model object returned by \code{estimate}}
}
\description{
This function is a wrapper for several small summary functions that provide
key summary statistics to help interpret the results of your model.

This function is a wrapper for several small summary functions that provide
key summary statistics to help interpret the results of your model.
}
