% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarizes outputs}
\usage{
summary(model, summary_opt)
}
\arguments{
\item{model}{A model opbject returned by \code{estimate}}

\item{summary_opt}{A list of summary options}
}
\description{
This function is a wrapper for several small summary functions that provide
key summary statistics to help interpret the results of your model.
}
