% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary_model_fit}
\alias{summary_model_fit}
\title{Summary of model fits}
\usage{
summary_model_fit(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}
}
\value{
A tibble
}
\description{
A wrapper function around \link{\code{model_fit}} returning all fit
calculations as a tibble.
}
