% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R, R/summary.R
\name{summary_parameter_info}
\alias{summary_parameter_info}
\title{Summary of parameter info}
\usage{
summary_parameter_info(object, ...)

summary_parameter_info(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Other arguments passed to the function}
}
\description{
Summary of parameter info

Summary of parameter info
}
