% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_info.R
\name{summary_worker_info}
\alias{summary_worker_info}
\title{Function for summarizing worker information}
\usage{
summary_worker_info(worker_info)
}
\arguments{
\item{worker_info}{A worker information object from get_worker_info()}
}
\description{
The function summarizes and prepares to print the worker information to a
.txt file
}
