% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{tidy.cmdlr}
\alias{tidy.cmdlr}
\title{Create a summary function for the coefficient matrix}
\usage{
\method{tidy}{cmdlr}(object, robust = TRUE, hypotheses = NULL, ...)
}
\description{
Create a summary function for the coefficient matrix
}
