% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidying the environment to prepare for estimation}
\usage{
tidy()
}
\description{
A small function to tidy and prepare the work space and environment to run
the \code{cmdlR} functions. It is highly recommended to call the function
directly using the \code{::} notation to avoid having to load packges twice.
}
