% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate model options}
\usage{
validate(model_options, db_names)
}
\arguments{
\item{model_options}{A user supplied list of model options}

\item{db_names}{A character string of variable names in the data}
}
\value{
A list of model options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of model options and sets 
defaults for missing options. All checks are intentionally aggressive
and will throw an error rather than make assumptions about what the user
intended to do.
}
\details{
The function is run once inside of \code{\link{prepare}} to correctly set up 
the estimation environment. It is meant for internal use only and is not 
exported.
}
