% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_estim_opt.R
\name{validate_estim_opt}
\alias{validate_estim_opt}
\title{Set default estimation options}
\usage{
validate_estim_opt(estim_opt_input)
}
\arguments{
\item{estim_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function sets default estimation options for options that are not 
specified by the user
}
