% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_save_opt}
\alias{validate_save_opt}
\title{Validates and sets default save options}
\usage{
validate_save_opt(input_save_opt)
}
\arguments{
\item{input_save_opt}{A named list of save options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of save options and sets 
defaults for the missing options.
}
\details{
The function is intended for internal use only.
}
