% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.cmdlr}
\alias{vcov.cmdlr}
\title{S3 generic for vcov}
\usage{
\method{vcov}{cmdlr}(object, robust = FALSE, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{robust}{A boolean equal to TRUE if you want to return a robust 
variance covariance matrix based on a simple sandwich estimator. The default
value is FALSE}

\item{...}{Other arguments passed to the function}
}
\value{
A matrix with row and column names equal to the parameters of the
fitted model
}
\description{
A generic method for obtaining standard and robust variance covariance matrix
}
