% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdlR.R
\docType{package}
\name{cmldR}
\alias{cmldR}
\alias{cmldR-package}
\title{cmdlR: Choice Modeling in R}
\description{
The problem of choice is fundamental to economics. Choice models are used to 
understand how people make choices in markets. cmdlR is an easy to use 
package for choice modeling in R. It contains several pre-programmed 
likelihood functions for the most common models and the framework is flexible 
so that the users can extend the functionality of the package with their own 
functions.
}
