% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work-in-progress.R
\name{convergence_criteria}
\alias{convergence_criteria}
\title{Calculate the convergence criteria}
\usage{
convergence_criteria(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}
}
\value{
A named numeric
}
\description{
Calculate the convergence criteria
}
