% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimates the model}
\usage{
estimate(log_lik, estim_opt, workers = NULL)
}
\arguments{
\item{log_lik}{A log likeliohood function to be maximized/minimized}

\item{estim_opt}{List of estimation options}

\item{workers}{A worker cluster for parallel estimation}
}
\description{
The function is a wrapper for several cmdlR functions that sets up parallel
processes and estimates the data.
}
