% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Function to estimate a choice model using cmdlr}
\usage{
estimate(ll, prepared_inputs, validated_options)
}
\arguments{
\item{ll}{This is the 'raw' log-likelihood function passed to the estimation
routine. It is important that the user takes into account whether the 
optimization routine is a minimizer (e.g. 'ucminf') or a maximizer 
(e.g. 'maxlik').}

\item{prepared_inputs}{A list of prepared inputs returned from 
\code{\link{prepare}}}

\item{validated_options}{A list of validated model options returned from 
\code{\link{validate}}}
}
\value{
A list with model results including the hessian, standard and
robust variance-covariance matrixes and
}
\description{
either maximize or minimize the log-likelihood function. If estimation 
is successful, the function calculates a high-precision numerical hessian
and a standard and robust variance-covariance matrix.
}
\details{
NOTE: There are no checks on whether the log-likelihood function is correctly
specified. If you see that your log-likelihood value is getting progressively
worse from starting values, check your 'll' specification and that it 
corresponds to the chosen optimizer.
}
\examples{
\dontrun{
  # See /examples for how to use
  estimate(ll, prepared_inputs, validated_options)
}

}
