% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_maxlik}
\alias{estimate_maxlik}
\title{Estimate the model using the maxLik package}
\usage{
estimate_maxlik(
  log_lik,
  param_free,
  param_fixed,
  workers,
  ll,
  return_sum,
  pb,
  estim_opt
)
}
\arguments{
\item{log_lik}{A log likelihood expression}

\item{param_free}{A named vector of freely estimated parameters}

\item{param_fixed}{A named vector of parameters to keep fixed at their
starting values during estimation}

\item{workers}{A list of workers or NA if single core}

\item{ll}{The 'raw' log-likelihood function}

\item{return_sum}{A boolean equal to TRUE if the log-likelihood function
shoudl return the sum of individual contributions. Default is FALSE.}

\item{pb}{A progress bar environment from the 'progress' package}

\item{estim_opt}{A list of estimation options}
}
\value{
A custom model object with the results fo the estimation.
}
\description{
The function is a convenient wrapper around \code{\link{maxLik}} and returns
a custom model object.
}
