% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-model.R
\name{estimate_ucminf}
\alias{estimate_ucminf}
\title{Estimate the model using the 'ucminf' package}
\usage{
estimate_ucminf(log_lik, param_free, param_fixed, workers, ll, control)
}
\arguments{
\item{log_lik}{A log-likelihood expression prepared by 
\code{\link{prepare_log_lik}}}

\item{param_free}{A list of freely estimated parameters}

\item{param_fixed}{A list of parameters that are held fixed at their 
starting values}

\item{workers}{A list of workers for parallel computing}

\item{ll}{The 'raw' log likelihood function passed by the user}

\item{control}{A list of estimation controls passed to the optimizer}
}
\value{
A custom model object with the results of the estimation
}
\description{
The function is a convenient wrapper around \code{\link{ucminf}}.
}
