% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{hessian}
\alias{hessian}
\title{Calculate the Hessian}
\usage{
hessian(type, log_lik, param_free, param_fixed, workers, ll, method)
}
\arguments{
\item{type}{A character string giving the method to use when approximating
the Hessian. One of: "numderiv" or "maxlik".}
}
\description{
The function is a wrapper around several methods for calculating the
numerical Hessian matrix. The function is meant for internal use and is not
exported.
}
