% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{hessian}
\alias{hessian}
\title{Calculate the Hessian}
\usage{
hessian(type, log_lik, param_free, param_fixed, workers, ll, method)
}
\arguments{
\item{type}{A character string giving the method to use when approximating
the Hessian. One of: "numderiv" or "maxlik".}

\item{log_lik}{A log-likelihood expression prepared by 
\code{\link{prepare_log_lik}}}

\item{param_free}{A list of freely estimated parameters}

\item{param_fixed}{A list of parameters that are held fixed at their 
starting values}

\item{workers}{A list of workers for parallel computing}

\item{ll}{The 'raw' log likelihood function passed by the user}

\item{method}{A method for the approximation of the Hessian. Only necessary
for the maxlik package}
}
\description{
The function is a wrapper around several methods for calculating the
numerical Hessian matrix. The function is meant for internal use and is not
exported.
}
