% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_name.R
\name{make_model_name}
\alias{make_model_name}
\title{A function to create a model name}
\usage{
make_model_name(model_name)
}
\arguments{
\item{model_name}{A string}
}
\description{
A function to create a model name
}
