% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_probs.R
\name{mnl_probs}
\alias{mnl_probs}
\title{Calculate the MNL probability}
\usage{
mnl_probs(V, model_opt)
}
\arguments{
\item{V}{A list of the observable part of utility}

\item{model_opt}{List of model options}
}
\description{
Function is convenient way to calculate the MNL probabilities.
}
\details{
Function for calculating the MNL probabilities

Calculate the MNL probabilities taking the difference from the chosen
alternative.
}
