% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nid}
\alias{nid}
\title{Function for extracting number of individuals}
\usage{
nid(object, ...)
}
\arguments{
\item{object}{A 'cmdlr' model object}

\item{...}{Additional parameters passed to the function}
}
\description{
A generic method for getting the number of individuals based on the lenght of
the vector of optimum values
}
