% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability-helpers.R
\name{panel_product}
\alias{panel_product}
\title{Take product over the sequence of choices}
\usage{
panel_product(x, n_ct)
}
\arguments{
\item{x}{A matrix with the probabilities of the chosen alternative}

\item{n_ct}{The number of choice situations per individual}
}
\value{
A vector of sequence probabilities. If our model contains inter-
individual draws, then this vector is n_id * n_draws
}
\description{
Calculates the product across choice observations for the same individual.
The function makes a call to \code{\link{reshape_panel}} and 
\code{\link{replace_na_padding}} prior to taking the
product.
}
