% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_product.R
\name{panel_product}
\alias{panel_product}
\title{Take the panel product}
\usage{
panel_product(probability, n_ct)
}
\arguments{
\item{probability}{The probability of the chosen alternative returned from
\code{mnl_probs}}

\item{n_ct}{The number of choie situations per individual}
}
\value{
A vector of sequence probabilities. If our model contains inter-
individual draws, then this vector is n_id * n_draws
}
\description{
Calculates the product across choice observations for the same individual
}
