% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{A function that prepares the data for estimation}
\usage{
prepare(opts, db)
}
\arguments{
\item{opts}{List of \code{estim_opt}, \code{model_opt}, \code{save_opt}.}

\item{db}{Data}
}
\value{
A list with \code{estim_opt}, \code{model_opt}, \code{save_opt},
\code{db} and \code{workers}.
}
\description{
The function is a wrapper for several cmdlR functions aimed at preparing the
data for estimation.
}
