% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{A function that prepares the data for estimation}
\usage{
prepare(db, lik, estim_opt, model_opt, save_opt, summary_opt)
}
\arguments{
\item{db}{Data¨}

\item{lik}{Likelihood function}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}

\item{save_opt}{List of options for saving outputs}

\item{summary_opt}{List of options for summary statistics}
}
\value{
A list with \code{estim_opt}, \code{model_opt}, \code{save_opt},
\code{db} and \code{workers}.
}
\description{
The function is a wrapper for several cmdlR functions aimed at preparing the
data for estimation.
}
