% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare the model components and log likelihood function for estimation}
\usage{
prepare(db, ll, validated_options)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}

\item{ll}{The log likelihood funciton}

\item{validated_options}{A list of options returned by \code{\link{validate}}}
}
\value{
A list with the estimation environment and the log likelihood
function
}
\description{
The function prepares the data, draws and alternative availabilites, and 
creates an estimation environment providing the context for evaluating the
log-likelihood function.
}
