% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare the model components and log likelihood function for estimation}
\usage{
prepare(
  ll,
  db,
  model_options,
  cores = 1,
  check_data = TRUE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{ll}{The log likelihood funciton}

\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}

\item{model_options}{A list of user supplied model options to be validated}

\item{cores}{An integer giving the number of cores to use in estimation. The
default is to use a single core.}

\item{check_data}{A boolean equal to TRUE if we should run data checks and 
pad the data with NA.}

\item{debug}{A boolean equal to TRUE if you are debugging the log likelihood
function. It forces the number of cores to 1. Then you can simply
\code{\link{attach}} the estimation environment and the list of parameters
to run the log likelihood function line by line. The default is FALSE.}
}
\value{
The estimation environment
}
\description{
The function prepares the data, draws and alternative availabilites, and 
creates an estimation environment providing the context for evaluating the
log-likelihood function.
}
