% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to check the data}
\usage{
prepare_data(db, estim_opt, model_opt)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} with the data to be used in
estimation.}

\item{estim_opt}{A list of options for the estimation procedure such as 
type of optimizer, approximation routines, tolerance levels etc.}

\item{model_opt}{A list of model options including random parameters, draws,
fixed and non-fixed parameters during estimation, starting value search 
procedures etc.}
}
\value{
A list of conditions
}
\description{
The function will check the data to see if it is balanced. It will take the
necessary steps to balance the data. If it is unable to do so, the code will
return suggestions for how to set up the data.
}
\details{
This function is intended for internal use only.
}
