% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare the data for estimation}
\usage{
prepare_data(db, str_id, str_ct, n_id, n_ct, cores, check_data)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} containing the data}

\item{str_id}{A character string giving the name of the id variable}

\item{str_ct}{A character string giving the name of the ct variable}

\item{n_id}{An integer giving the nubmer of respondents}

\item{n_ct}{An integer giving the number of choice situations}

\item{cores}{An integer indicating the number of cores to use in estimation}

\item{check_data}{A boolean eqaul to TRUE if the code should skip the 
data checks on number of choice tasks per individual.}
}
\value{
A \code{\link{data.frame}} padded with NA
}
\description{
The function checks that the data is supplied as a data.frame or tibble and
that it contains an equal number of choice observations per individual.
}
\details{
NOTE: While padding allows the use of fast matrix operations, it may come
at the cost of disproportionate memory use for large datasets with many
observations and variables.
}
