% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_draws.R
\name{prepare_draws}
\alias{prepare_draws}
\title{Prepares the draws}
\usage{
prepare_draws(db, estim_opt, model_opt)
}
\arguments{
\item{db}{Data}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}
}
\description{
A function that will create and prepare the draws for estimation
}
