% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_log_lik.R
\name{prepare_log_lik}
\alias{prepare_log_lik}
\title{Prepares the log likelihood function}
\usage{
prepare_log_lik(ll, estim_env, model_opt, workers)
}
\arguments{
\item{ll}{A log likelihood function. If \code{estim_opt$optimizer = 'ucminf'},
then the log likelihood function should return the negative log likelihood
value.}

\item{estim_env}{A list of data and other variables to be coerced to environment with eval}

\item{model_opt}{A list of model options including random parameters, draws,
fixed and non-fixed parameters during estimation, starting value search 
procedures etc.}

\item{workers}{A PSOCK or Fork cluster of workers created using 
\code{parallel::makeCluster()}}
}
\value{
The estimation ready log likelihood function
}
\description{
The function takes the user supplied log-likelihood function and creates
convenient wrappers that ensure the function can be evaluated both using a
single core and multiple cores.
}
\details{
The function is meant for internal use only.
}
