% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_log_lik}
\alias{prepare_log_lik}
\title{Prepares the log likelihood function}
\usage{
prepare_log_lik(ll, estim_env, workers)
}
\arguments{
\item{ll}{The log likelihood funciton}

\item{estim_env}{An estimation environment returned by 
\code{\link{prepare_estimation_environment}}}

\item{workers}{A list of workers created using the parallel package}
}
\value{
The estimation-ready log likelihood function
}
\description{
The function takes the user supplied log-likelihood function and creates
convenient wrappers that ensure the function can be evaluated both using a
single core and multiple cores.
}
