% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_log_lik.R
\name{prepare_log_lik}
\alias{prepare_log_lik}
\title{Prepares the log likelihood function}
\usage{
prepare_log_lik(log_lik, estim_opt, model_opt, workers)
}
\arguments{
\item{log_lik}{Log likelihood function}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}

\item{workers}{A PSOCK cluster of workers}
}
\value{
A log likelihood wrapper functions
}
\description{
This function puts the finishing touches on the user supplied log likelihood
function and ensures that it can be run on a single core or multiple cores
using both the 'maxlik' and 'nloptr' packages.
}
