% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_num_grad.R
\name{prepare_num_grad}
\alias{prepare_num_grad}
\title{Prepares the numerical gradient}
\usage{
prepare_num_grad()
}
\value{
A numerical gradient wrapper function
}
\description{
'nloptr' and 'trustOptim' requires a gradient. This function is a wrapper
for numDeriv::grad() and prepares a high-precision numerical gradient to
use with 'nloptr' and 'trustOptim'.
}
