% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_num_grad}
\alias{prepare_num_grad}
\title{Prepares the numerical gradient}
\usage{
prepare_num_grad(ll, estim_env, workers)
}
\arguments{
\item{ll}{This is the 'raw' log-likelihood function passed to the estimation
routine. It is important that the user takes into account whether the 
optimization routine is a minimizer (e.g. 'ucminf') or a maximizer 
(e.g. 'maxlik').}

\item{estim_env}{An estimation environment returned by 
\code{\link{prepare_estimation_environment}}}

\item{workers}{A list of workers created using the parallel package}
}
\value{
A high precision numerical gradient function
}
\description{
The optimization routines 'ucminf', 'nloptr' and 'trustOptim' requires the
user to supply a gradient. Writing an analytical gradient can be quite
cumbersome for very complex likelihood expressions. This function is a simple
wrapper around \code{numDeriv::grad()} and prepares a high-precision 
numerical gradient that can be supplied directly to the optimizers that 
require one. 
Note that a numerical gradient is slower in calculation and less precise than 
an analytical gradient.
}
