% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_workers.R
\name{prepare_workers}
\alias{prepare_workers}
\title{Function to prepare the workers for parallel estimation}
\usage{
prepare_workers(db, estim_opt, model_opt, save_opt)
}
\arguments{
\item{db}{Data}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}

\item{save_opt}{List of options for saving outputs}
}
\value{
A cluster of workers
}
\description{
The function will set up the workers, split the data and export necessary
objects and functions to the workers.
}
