% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_workers.R
\name{prepare_workers}
\alias{prepare_workers}
\title{Function preparing the workers}
\usage{
prepare_workers(db, draws, workers, estim_opt, model_opt, save_opt)
}
\arguments{
\item{db}{A \code{data.frame()} or \code{tibble()} with the data to be used in
estimation.}

\item{draws}{A list of draws where each list element correspond to one dimension
or if we are using multiple cores, a list of lists where the top-level list
corresponds to the cores and the lower-level list corresponds to the draws.}

\item{workers}{A PSOCK or Fork cluster of workers created using 
\code{parallel::makeCluster()}}

\item{estim_opt}{A list of options for the estimation procedure such as 
type of optimizer, approximation routines, tolerance levels etc.}

\item{model_opt}{A list of model options including random parameters, draws,
fixed and non-fixed parameters during estimation, starting value search 
procedures etc.}

\item{save_opt}{List of options for storing the outputs including which
model outputs to store, how to store them and where to store them.}
}
\value{
The function does not return any objects.
}
\description{
When we estimate the model using multiple cores, we need to set up the workers
such that the correctly split data, draws, packages and settings are available
to the log-likelihood function.
}
\details{
The function is meant for internal use only.
}
