% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{robust_vcov}
\alias{robust_vcov}
\title{A sandwich estimator for robust standard errors}
\usage{
robust_vcov(object, ...)
}
\value{
A matrix with row and column names equal to the parameters of the
fitted model
}
\description{
A straight forward sandwich estimator for robust standard errors with a 
correction based on the number of individuals/respondents in the data. The
variance-covariance matrix is based on the 'cmdlr' \link{\code{vcov}} 
function.
}
