% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_start_values.R
\name{search_start_values}
\alias{search_start_values}
\title{Search for starting values}
\usage{
search_start_values(
  ll,
  estim_env,
  model_options,
  control = NULL,
  type = "simple",
  n_candidates = 100,
  n_return = 10,
  multiplier = 1
)
}
\arguments{
\item{ll}{This is the 'raw' log-likelihood function passed to the estimation
routine. It is important that the user takes into account whether the 
optimization routine is a minimizer (e.g. 'ucminf') or a maximizer 
(e.g. 'maxlik').}

\item{estim_env}{An estimation environment or list of estimation environments
returned from \code{\link{prepare}}}

\item{model_options}{A list of model options. Note, this list is validated
for a second time here to set some necessary defaults. See
\code{\link{validate}} for details.}

\item{control}{A list of control options that are passed to
\code{\link{set_controls}}.}

\item{type}{A character string indicating whether to use a simple or a smart
search algorithm. Default is 'simple'}

\item{n_candidates}{An integer giving the number of candidates to evaluate.
Default is 100.}

\item{n_return}{An integer giving the number of parameter vectors to return.
The default is 10}

\item{multiplier}{A double indicating a multiplier for the 'simple' search
algorithm. The default is 1.}
}
\value{
A matrix of starting values
}
\description{
The function searches for new starting values. Either in a random fashion or
based on a smarter search algorithm. The smart search algorithm is a rewritten
version of the search algorithm implemented in the Apollo package. The function
is rewritten to work with the current modeling framework.
}
\references{
Hess, S. & Palma, D., 2019, Apollo: A flexible, powerful and customisable f
reeware package for choice model estimation and application, Journal of 
Choice Modelling, 32
Bierlaire, M., Thémans, M. & Zufferey, N., 2010, A heuristic for nonlinear 
global optimization, INFORMS Journal on Computing, 22(1):
}
