% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_start_values.R
\name{search_start_values}
\alias{search_start_values}
\title{Function to search for starting values}
\usage{
search_start_values(inputs)
}
\arguments{
\item{inputs}{List of inputs from}
}
\value{
A list of parameters
}
\description{
The function searches for new starting values. Either in a random fashion or
based on a smarter search algorithm. The smart search algorithm is a rewritten
version of the search algorithm implemented in the Apollo package. The function
is rewritten to work with the current modelign framework.

Suggested usage:
  Direct assignment to inputs$model_opt$param
}
\references{
Hess, S. & Palma, D., 2019, Apollo: A flexible, powerful and customisable freeware package for choice model estimation and application, Journal of Choice Modelling, 32
Bierlaire, M., Thémans, M. & Zufferey, N., 2010, A heuristic for nonlinear global optimization, INFORMS Journal on Computing, 22(1):
}
