% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_estim_opt_defaults.R
\name{set_estim_opt_defaults}
\alias{set_estim_opt_defaults}
\title{Set default estimation options}
\usage{
set_estim_opt_defaults(estim_opt_input)
}
\arguments{
\item{estim_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function sets default estimation options for options that are not 
specified by the user
}
