% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Function to split the data}
\usage{
split_data(db, estim_opt, model_opt)
}
\arguments{
\item{db}{A data frame}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}
}
\value{
A list of data matrices with length equal to \code{cores}.
}
\description{
The function takes the data and splits it into roughly equal chunks to be
sent to the workers for parallel processing. The function takes care to not
split the same individual across multiple workers.
}
\details{
It is important that the function is run after \code{check_data}. The code
only splits individuals and observations correctly if there are equal
number of rows per individual.
}
