% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarizes outputs}
\usage{
summarize(model, summary_opt)
}
\arguments{
\item{model}{A model opbject returned by \code{estimate}}

\item{summary_opt}{A list of summary options}

\item{model_opt}{A list of model options}
}
\description{
This function is a wrapper for several small summary functions that provide
key summary statistics to help interpret the results of your model.
}
