% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cmdlr}
\alias{summary.cmdlr}
\title{S3 Generic for model summary}
\usage{
\method{summary}{cmdlr}(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Other arguments passed to the function}
}
\description{
S3 Generic for model summary
}
