% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom-generics.R
\name{tidy.cmdlr}
\alias{tidy.cmdlr}
\title{Tidy a cmdlr object}
\usage{
\method{tidy}{cmdlr}(x, robust = TRUE, hypotheses = NULL, ...)
}
\arguments{
\item{x}{A model object of class 'cmdlr'}

\item{robust}{A boolean equal to TRUE if robust standard errors should be 
used and reported. The default value is TRUE.}

\item{hypotheses}{A vector the length of the number of parameters (incl.
free and fixed) with values for the null hypothesis to test against. The
default is a two-sided test against 0.}

\item{...}{Additional parameters passed to tidy}
}
\value{
A tidy [tibble::tibble()] summarizing the component-level information
about the model
}
\description{
Tidy a cmdlr object
}
