% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validates inputs}
\usage{
validate(log_lik, estim_opt, model_opt, save_opt, summary_opt)
}
\arguments{
\item{log_lik}{Log likelihood function}

\item{estim_opt}{List of estimation options}

\item{model_opt}{List of model options}

\item{save_opt}{List of options for saving outputs}

\item{summary_opt}{List of options for summary statistics}
}
\value{
A list with updated and validated \code{estim_opt}, \code{model_opt},
and \code{save_opt}.
}
\description{
The function is a wrapper for several cmdlR functions that validates the
inputs. The purpose is to ensure correct specification of the model and that
all information is available.
}
