% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validates the inputs to the model estimation}
\usage{
validate(estim_opt)
}
\arguments{
\item{estim_opt}{List of estimation options}
}
\description{
The function is a wrapper for several cmdlR functions that checks the inputs,
controls and data. The purpose is to make sure that the combination of 
inputs are allowed and supported by the package.
}
