% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate all options}
\usage{
validate(input_estim_opt, input_model_opt, input_save_opt, db)
}
\arguments{
\item{input_estim_opt}{A named list of estimation options}

\item{input_model_opt}{A named list of model options}

\item{input_save_opt}{A named list of save options}

\item{db}{A data.frame with the data used for estimation}
}
\value{
Returns a list with three elements.
}
\description{
The function validates all user supplied lists of options and sets reasonable
default values for missing options. All checks are intentionally aggressive
and will throw an error rather than make assumptions about what the user
intended to do.
}
