% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_estim_opt.R
\name{validate_estim_opt}
\alias{validate_estim_opt}
\title{Validates and sets default estimation options}
\usage{
validate_estim_opt(estim_opt_input)
}
\arguments{
\item{estim_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of estimation options and sets 
defaults for missing options. The checks are aggressive and will stop
the estimation and ask the user to respecify incorrect inputs. This is to 
ensure that the user intends to use these exact options.
}
\details{
The function is intended for internal use only.
}
