% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_estim_opt}
\alias{validate_estim_opt}
\title{Validates and sets default estimation options}
\usage{
validate_estim_opt(input_estim_opt)
}
\arguments{
\item{input_estim_opt}{A named list of estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of estimation options and sets 
defaults for missing options.
}
