% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_estim_opt.R
\name{validate_estim_opt}
\alias{validate_estim_opt}
\title{Validates and sets default estimation options}
\usage{
validate_estim_opt(estim_opt_input)
}
\arguments{
\item{estim_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of estimation options and sets 
defaults for missing options. If these options are essential, an error 
will occur and the program will stop. If the options are important, a
warning will be posted to the terminal and a default option will be used, 
but the program will continue.
}
