% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_model_opt.R
\name{validate_model_opt}
\alias{validate_model_opt}
\title{Set default model options}
\usage{
validate_model_opt(model_opt_input)
}
\arguments{
\item{model_opt_input}{A list of user specified estimation options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function sets default model options for options that are not 
specified by the user
}
