% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_save_opt.R
\name{validate_save_opt}
\alias{validate_save_opt}
\title{Validates and sets default estimation options}
\usage{
validate_save_opt(save_opt_input)
}
\arguments{
\item{save_opt_input}{A list of user specified save options}
}
\value{
Returns a list of options with missing input values replaced by 
default values
}
\description{
The function validates the user supplied list of save options and sets 
defaults for the missing options. If these options are essential, an error 
will occur and the program will stop. If the options are important, a
warning will be posted to the terminal and a default option will be used, 
but the program will continue.
}
\details{
\describe{
  \item{path}{The path and file name for your output. Defaults to current
  working directory and 'model-01'}
  \item{save_summary}{If \code{TRUE} save summary to .txt file specified
  in \code{path}. Default is \code{FALSE}.}
  \item{save_model_object}{If \code{TRUE} save model object to .rds file
  specified in \code{path}. Default is \code{FALSE}.}
  \item{save_hessian}{If \code{TRUE} save hessian to .csv file specified in
  \code{path}. Default is \code{FALSE}.}
  \item{save_conditionals}{If \code{TRUE} save the conditional estimates to
  a .csv file specified in \code{path}. Only valid for models with discrete
  or continuous mixing distributions. Default is \code{FALSE}.}
  \item{save_worker_info}{If \code{TRUE} save information about workers in
  .txt file specified in \code{path}. Only valid if \code{cores > 1}.
   Default is \code{FALSE}.}
}

The function is intended for internal use only.
}
