% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.cmdlr}
\alias{vcov.cmdlr}
\title{S3 generic for vcov}
\usage{
\method{vcov}{cmdlr}(object, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{...}{Other arguments passed to the function}
}
\value{
A matrix with row and column names equal to the parameters of the
fitted model
}
\description{
A generic method for obtaining the variance covariance matrix. It is a 
wrapper around the \link{\code{ginv}} from the MASS package to calculate the
(negative) inverse of the Hessian matrix.
}
