% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.cmdlr}
\alias{vcov.cmdlr}
\title{S3 generic for vcov}
\usage{
\method{vcov}{cmdlr}(object, robust = FALSE, ...)
}
\arguments{
\item{object}{A model object of class 'cmdlr'}

\item{robust}{If TRUE return the robust vcov}

\item{...}{Other arguments passed to the function}
}
\description{
S3 generic for vcov
}
